// ui/packages/src/api/generated/filesys/schemas/permanentlyDeleteRecycleCommand.ts
/**
 * Generated by orval v6.31.0 🍺
 * Do not edit manually.
 * Refactored free-fs with a subject-match-java style layered architecture.
 * OpenAPI spec version: v1.0.0
 */

export interface PermanentlyDeleteRecycleCommand {
  /** No comments found. */
  fileIds: string[];
  /** No comments found. */
  userId: string;
}